drivers: refactor to expose generic driver interfaces.
